backend/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
Pillow
numpy
moviepy
imageio-ffmpeg
python-multipart
